Add an experimental idea for general juxtaposition
Instead of just allowing S [] as a single-purpose type operator, we could
allow square bracketed types as the RHS. We cannot allow general type
juxtaposition (which would be useful for expressing prefix type operators)
because it will be ambiguous.

For example, S {} is a problem if used in a function return type position
as it's amibguous with the function body itself.